move logic to subscribers to centralize it
