+++ tests/publish/get_queue_items_tests.py
@@ -19,6 +19,7 @@ from superdesk.utc import utcnow
 from apps.publish.enqueue import enqueue_service
 from superdesk.publish.publish_queue import PUBLISHED_IN_PACKAGE
 from superdesk.publish import publish_queue
+from superdesk.publish.subscribers import SubscribersService
 from superdesk.metadata.item import CONTENT_TYPE, ITEM_TYPE
 
 
@@ -178,28 +179,32 @@ class QueueItemsTestCase(TestCase):
     @mock.patch.object(publish_queue, "get_resource_service")
     def test_on_create_hydrates_missing_destination_secrets(self, fake_get_resource_service):
         subscriber_id = ObjectId()
-        subscriber_service = fake_get_resource_service.return_value
-        subscriber_service.find_one.return_value = {
-            "_id": subscriber_id,
-            "destinations": [
-                {
-                    "_id": "ftp-destination-1",
-                    "name": "FTP Destination",
-                    "format": "ftp ninjs",
-                    "delivery_type": "ftp",
-                    "config": {
-                        "host": "127.0.0.1",
-                        "username": "superdesk",
-                        "password": "superdesk",
-                        "secret_token": "secret-token",
-                        "apiKey": "api-key",
-                        "access_key_id": "access-key-id",
-                        "secret_access_key": "secret-access-key",
-                        "passive": True,
-                    },
-                }
-            ],
-        }
+        subscriber_service = SubscribersService("subscribers", backend=MagicMock())
+        subscriber_service.find_one = MagicMock(
+            return_value={
+                "_id": subscriber_id,
+                "destinations": [
+                    {
+                        "_id": "ftp-destination-1",
+                        "name": "FTP Destination",
+                        "format": "ftp ninjs",
+                        "delivery_type": "ftp",
+                        "config": {
+                            "host": "127.0.0.1",
+                            "username": "superdesk",
+                            "password": "superdesk",
+                            "secret_token": "secret-token",
+                            "apiKey": "api-key",
+                            "access_key_id": "access-key-id",
+                            "secret_access_key": "secret-access-key",
+                            "passive": True,
+                        },
+                    }
+                ],
+            }
+        )
+        subscriber_service.generate_sequence_number = MagicMock(return_value=7)
+        fake_get_resource_service.return_value = subscriber_service
 
         service = publish_queue.PublishQueueService(backend=MagicMock())
         doc = {
@@ -214,7 +219,6 @@ class QueueItemsTestCase(TestCase):
                     "passive": True,
                 },
             },
-            "published_seq_num": 2,
             "state": "pending",
         }
 
@@ -225,3 +229,4 @@ class QueueItemsTestCase(TestCase):
         self.assertEqual(doc["destination"]["config"]["apiKey"], "api-key")
         self.assertEqual(doc["destination"]["config"]["access_key_id"], "access-key-id")
         self.assertEqual(doc["destination"]["config"]["secret_access_key"], "secret-access-key")
+        self.assertEqual(doc["published_seq_num"], 7)
